fix: green the integration test suite
Three IT failures surfaced when the uncommitted work first hit CI:
- LegislativeItemImporter.runWindow re-scanned the (date-filter-ignoring) drafts
  catalogue once per 7-day sub-window, double-counting upserts. It now does a
  single pass over the whole range, which also removes the duplicate list scan.
- The retention repo's bulk delete declared a `long` return, which @Modifying
  rejects at execution time; changed to `int` (+ test).
- Updated the vote-importer test to assert the improved behaviour: a failed
  detail fetch now rolls the record back entirely (no voteless event) and the
  run finishes PARTIAL instead of silently SUCCESS.

+++ backend/src/main/java/com/politico/source/SourceSnapshotRepository.java
@@ -40,7 +40,7 @@ public interface SourceSnapshotRepository extends JpaRepository<SourceSnapshot, 
     @Transactional
     @Query("delete from SourceSnapshot s "
             + "where s.entityType in :entityTypes and s.fetchedAt < :cutoff")
-    long deleteByEntityTypeInAndFetchedAtBefore(
+    int deleteByEntityTypeInAndFetchedAtBefore(
             @org.springframework.data.repository.query.Param("entityTypes")
                     Collection<String> entityTypes,
             @org.springframework.data.repository.query.Param("cutoff") Instant cutoff
